# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School API",
        "description": "API for managing school data",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/api/v1/students/{student}/result-pins": {
            "get": {
                "tags": [
                    "school-v1.4"
                ],
                "summary": "List result PINs for a student",
                "description": "Returns a student's result PINs filtered by session and term.",
                "operationId": "e90a79bf25c3667c25dfdc6b211ef80b",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "session_id",
                        "in": "query",
                        "description": "Session ID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "term_id",
                        "in": "query",
                        "description": "Term ID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List returned"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "school-v1.4"
                ],
                "summary": "Generate a result PIN for a student",
                "operationId": "b39cca4cc635a929bd72660612e1a176",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id",
                                    "term_id"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "term_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "expires_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-31T23:59:59Z"
                                    },
                                    "regenerate": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "max_usage": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PIN generated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/result-pins/bulk": {
            "post": {
                "tags": [
                    "school-v1.4"
                ],
                "summary": "Bulk-generate result PINs",
                "description": "Generate result PINs for many students using class or explicit student filters.",
                "operationId": "a4b7cf7ac22f588f9703d088413ce893",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id",
                                    "term_id"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "term_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "school_class_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "class_arm_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "student_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    },
                                    "regenerate": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "expires_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-31T23:59:59Z"
                                    },
                                    "max_usage": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bulk generation completed"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/result-pins": {
            "get": {
                "tags": [
                    "school-v1.4"
                ],
                "summary": "List result PINs across the school",
                "description": "Requires session_id and term_id; supports filtering by student, class, arm, and status.",
                "operationId": "bf2c9ea68d9a145b565949d1fc85d3d3",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "term_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "school_class_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_arm_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List returned"
                    },
                    "422": {
                        "description": "Missing required filters"
                    }
                }
            }
        },
        "/api/v1/result-pins/{resultPin}/invalidate": {
            "put": {
                "tags": [
                    "school-v1.4"
                ],
                "summary": "Invalidate a result PIN",
                "operationId": "aaba4276592d351c65c53e1b6a387ef7",
                "parameters": [
                    {
                        "name": "resultPin",
                        "in": "path",
                        "description": "Result PIN ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PIN invalidated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}